Introduced GameAction
GameAction is an abstract class that represents
actions that the user can execute on Klondike

So far, only StockAction is implemented (move the stock)

The GameActionController visits the GameAction and
acts properly. GameActionController knows Klondike.

SelectActionView prompts the user to select an action.
It is instantiated from KlondikeConsoleView.

Removed StockController. The actions on the stock
are simple enough to not need a specific controller
for them.