Refactor a bit...
The aim is to be able to let MoveScoreCalculator
be a TableauPile observer to increase the score
when turning up a card.

Need to increase the scope of GameActionScoreCalculator
and MoveScoreCalculator. Now they are class members...
well, there are shared pointers, we want disposable
instances.

MoveScoreCalculator holds now a reference to the scoreDelta
of GameActionScoreCalculator, so we do not need a getter
there.

Changed how GameActionController interacts with the ScoreController.
Before doing the action, the ScoreController will visit the
action, but won't do any score update. Then action will be
done, and then score will be updated