Fix some magic numbers
The Deck will know how many suits and how
many cards per suit it has, rather than
an absolute number of cards.

Foundations have a parameter: number of cards
so they can be considered complete. it will
be set to number of cards per suit, but to
test easier let's keep it at 2