Refactor: Define GameAction interface...
...and align StockAction and Move to it.

Main consequence is that StockAction now
has logic to actually do the action.
Hence, it needs a reference to the Stock,
passed via constructor

Since both kind of actions are now aligned
to the interface, there is no need now
for GameActionController to be a GameActionVistitor.
It will just call GameAction::doAction