Refactor Stock
Created aux functions to move cards between piles.

Implemented the stock movement in a generic way so it
can be reused when moving forward or backward (when undoing
a stock action).

Public functions are:
moveForward: from covered cards to waste, up to three cards. if there
are not covered cards, put all cards in waste back to covered.
Returns number of cards moved.

moveBackward: from waste to covered, up to the number of cards
specified as a parameter. If there are not cards in waste,
put all covered cards in waste, showing the last one.