Undo the score
After doing the action, the GameActionController
sets the scoreDelta of the action. The
GameActionScoreCalculator, when receiving an UndoAction,
gets the score awarded to the action that is being
undone, and multiplies it by -1. Then the score
will be updated with this inverse delta, effectively
undoing the score awarded by the action undone