Changed Move to be abstract
Three kinds of Move:
 - MoveFromStock
 - MoveFromTableauPile
 - MoveFromFoundation

Added MoveVisitor, to differenciate between
these kinds of movements

Move does not hold a pointer to a MoveOrigin.
It will be the subclass (MoveFromXX) the one
holding a pointer to the specific origin
(Stock, TableauPile or Foundation)

GameActionScoreCalculator is not a MoveOriginVisitor
now, but a MoveVisitor

RandomGameActionController now needs to instantiate
specific kinds of Move depending on what is the
origin, so it can't use the list of all MoveOrigins

MoveCardView is now a MoveVisitor which instantiates
a specific move depending on the user input
for move origin.