Refactored MoveDest
Exctracted common reusable logic for all
MoveDests. Dests only have two virtual functions:
 - contidion to add the first card
 - condition to add subsequent cards, as two argument
   function where one is the reference card and the second
   is the card to be added.

All dests have a pile of faceUpCards.

Common function to check if a pile can be added. It uses
Pile::isOrdered passing the condition to add subsequent
card as a criteria.

Common function to check if a card can be added. it will
check condition to add first card or to add subsequent cards.

Cards are added one by one, asserting that a card can be added
for every card